YTI-149 - All existing ES client code was removed and refactored to use Spring RestTemplate instead.
As a part of the task YTI-149 the "old" (ie. 5.2.2 version) of esRestClient was removed in order to be able to use the new High Level client (version 6.2.2) from now on. There seemed to be a jar hell as a result of using the old client and the new High Level Client simultaneously, so the code using the old client was refactored to use Spring RestTemplate instead.