extend terminology search query to organization label
Query string will now apply to organization label as well, in addition
to the terminology label.

Search by organization IDs was added in commit 3519c65,
but it was then determined that the preferred approach would be to
search the organization name instead. The previous implementation is
left as-is, although it might no be used.

YTI-1582